Go SQL driver: Renamed package from "client" to "vitessdriver".
This is more explicit. The Go "sql" package also talks about "drivers": https://golang.org/pkg/database/sql/

To get a sense how other drivers are named, here's a list: https://github.com/golang/go/wiki/SQLDrivers

We decided to have "vitess" in the name as well to make it clear in application code that this is the Vitess Go SQL driver.

also: Renamed main files from client*.go to driver*.go.
