+++ go/vt/vitessdriver/fakeserver_test.go
@@ -1,4 +1,4 @@
-package client
+package vitessdriver
 
 import (
 	"errors"
