+++ go/vt/vitessdriver/plugin_gorpcvtgateconn.go
@@ -2,12 +2,12 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package client
+package vitessdriver
 
-// Imports and register the gRPC vtgateconn client.
+// Imports and register the gorpc vtgateconn client.
 
 import (
-	// Import the gRPC vtgateconn client and make it part of the Vitess
+	// Import the gorpc vtgateconn client and make it part of the Vitess
 	// Go SQL driver. This way, users do not have to do this.
-	_ "github.com/youtube/vitess/go/vt/vtgate/grpcvtgateconn"
+	_ "github.com/youtube/vitess/go/vt/vtgate/gorpcvtgateconn"
 )
